refactor(privacy): share notice-seen mutation; fix marker return type; harden flaky e2e
- Extract useMarkPrivacyNoticeSeen() so the consent modal and onboarding share
  one markNoticeSeen + gate-cache-update path (addresses cubic P3: two drifting
  implementations).
- Widen markNoticeSeen IPC return type to Result<{ privacy_notice_seen: boolean }>
  to match the backend response (cubic P3).
- Bump the privacy-consent.t2 waitForConsentModal poll to 20s; the default 5s
  raced app-launch + config migration on loaded CI runners (flaky, green on retry).
